requirements.txt
torch~=1.10.2
wandb~=0.12.11
Pillow~=9.0.1
torchvision~=0.11.3
fastai~=2.5.4